% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_bg.R
\name{c19_bg_data}
\alias{c19_bg_data}
\title{Provide access to Bulgarian COVID-19 data from data.egov.bg.}
\usage{
c19_bg_data(reload = FALSE, redownload = FALSE)
}
\arguments{
\item{reload}{reload from disk}

\item{redownload}{refresh from internet}
}
\description{
Provide access to Bulgarian COVID-19 data from data.egov.bg.
}
\seealso{
Other data funcs: 
\code{\link{c19_eu_data}()},
\code{\link{c19_nsi_data}()},
\code{\link{c19_reload}()}
}
\concept{data funcs}
