% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_eu_plots.R
\name{c19_deaths_factor}
\alias{c19_deaths_factor}
\title{Plot excess deaths factors for a selection of countries.}
\usage{
c19_deaths_factor(
  countries = c("BG", "UK", "BE", "NL", "FR", "ES", "IT", "RO"),
  eu_data = c19_eu_data()
)
}
\arguments{
\item{countries}{countries to plot}

\item{eu_data}{eu data}
}
\description{
Plot excess deaths factors for a selection of countries.
}
\seealso{
Other plot funcs: 
\code{\link{c19_deaths_age}()},
\code{\link{c19_deaths_map}()},
\code{\link{c19_deaths_total}()},
\code{\link{c19_eu_weekly}()},
\code{\link{c19_heat}()},
\code{\link{c19_oblasts}()},
\code{\link{c19_r_plot}()},
\code{\link{c19_var_plot}()}
}
\concept{plot funcs}
