% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_eu_plots.R
\name{c19_deaths_total}
\alias{c19_deaths_total}
\title{Plot total weekly deaths for a country.}
\usage{
c19_deaths_total(country_code, eu_data = c19_eu_data())
}
\arguments{
\item{country_code}{two-letter EU code e.g. "BG", "UK", "EL" (=Greece)}

\item{eu_data}{eu data}
}
\description{
Plot total weekly deaths for a country.
}
\seealso{
Other plot funcs: 
\code{\link{c19_deaths_age}()},
\code{\link{c19_deaths_factor}()},
\code{\link{c19_deaths_map}()},
\code{\link{c19_eu_weekly}()},
\code{\link{c19_heat}()},
\code{\link{c19_oblasts}()},
\code{\link{c19_r_plot}()},
\code{\link{c19_var_plot}()}
}
\concept{plot funcs}
