% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_estimate_r.R
\name{c19_estimate_r}
\alias{c19_estimate_r}
\title{Estimate R values.}
\usage{
c19_estimate_r(country_data = c19_bg_data())
}
\arguments{
\item{country_data}{test country data (default: c19_bg_data())}
}
\description{
Estimate R values.
}
