% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_eu.R
\name{c19_eu_data}
\alias{c19_eu_data}
\title{Provides access to ECDC/EUROSTAT COVID-19 data.}
\usage{
c19_eu_data(reload = FALSE, redownload = FALSE)

c19_eu_data(reload = FALSE, redownload = FALSE)
}
\arguments{
\item{reload}{reload from disk}

\item{redownload}{refresh from internet}
}
\description{
Provides access to ECDC/EUROSTAT COVID-19 data.

Provide access to ECDC/EUROSTAT COVID-19 data.
}
\seealso{
Other data funcs: 
\code{\link{c19_bg_data}()},
\code{\link{c19_nsi_data}()},
\code{\link{c19_reload}()}

Other data funcs: 
\code{\link{c19_bg_data}()},
\code{\link{c19_nsi_data}()},
\code{\link{c19_reload}()}
}
\concept{data funcs}
