% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_eu_plots.R
\name{c19_eu_weekly}
\alias{c19_eu_weekly}
\title{Weekly plot from ECDC/EUROSTAT data}
\usage{
c19_eu_weekly(
  indicator,
  continents = c("Asia", "Africa", "Europe", "Oceania", "America"),
  top_n = 20,
  lower_x = 10,
  lower_y = NA,
  label_fun = round,
  axis_labels = scales::label_number(),
  eu_data = c19_eu_data()
)
}
\arguments{
\item{indicator}{one of: "r14_cases", "r14_deaths", "em_1m", hosp_1m",
"positivity", "tests_100k"}

\item{continents}{def = c("Asia", "Africa", "Europe", "Oceania", "America")}

\item{top_n}{number of lines to label (default 20)}

\item{lower_x}{week axis limit (default: NA = show all data)}

\item{lower_y}{indicator axis limit (default: NA = show all data)}

\item{label_fun}{function to apply to line labels (default = round)}

\item{axis_labels}{how to label axis (default: scales::label_number())}

\item{eu_data}{eu data (default: c19_eu_data())}
}
\value{
A ggplot
}
\description{
Weekly plot from ECDC/EUROSTAT data
}
\examples{
c19_eu_weekly(indicator = "r14_cases", lower_y = 0)
c19_eu_weekly(indicator = "positivity",
              top_n = 100,
              label_fun = function(x) sprintf("\%.1f\%\%", 100 * x),
              axis_labels = scales::label_percent())
}
