% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_inkmagick.R
\name{c19_inkmagick}
\alias{c19_inkmagick}
\title{Produces high-quality PNG & JPEG}
\usage{
c19_inkmagick(var = TRUE, eu = TRUE, r = TRUE, d_all = FALSE)
}
\arguments{
\item{var}{whether to output various data.egov.bg-sourced plots}

\item{eu}{whether to output EUROSTAT/ECDC-sourced plots}

\item{r}{whether to plot r (incl. time-consuming calculation)}

\item{d_all}{whether to plot age band death plots for countries other than BG}
}
\description{
Production quality output using Inkscape for rasterization & ImageMagick
for JPEG compression. (SVG output is the same quality as other functions.)
}
\details{
Using these programs yields slightly better results but is not really
necessary. c19_save_all() saves SVG at the same quality but produces
lower-quality PNG/JPEG. You can also save individual plots using
ggplot2::ggsave() or plot to screen (e.g. [c19_heat()] or [print(c19_heat())]
in non-interactive mode).

Output files will be stored in folders c19bg/plots/MMMDD (month-day).
Downloaded and calculated data will be stored in c19bg/data.

Paths to Inkscape and ImageMagick can be set via options.
}
\examples{
\dontrun{
# set options
options(c19bg.output = list(
  inkscape = "\"C:\\\\Program Files\\\\Inkscape\\\\bin\\\\inkscape.exe\"",
  inkopts = "-w \%d --export-filename",
  magick = "magick",  # works, if it's in PATH
  imgkopts = "-quality 100",
  pixwidth = 1375,
  width = 11,
  height = 7,
))
options(c19bg.output_dir = "c19bg/plots")

# standard run
c19_inkmagick()

# include country-level age band plots from EUROSTAT demo mortality database
c19_inkmagick(d_all = TRUE)
}
}
\seealso{
Other output funcs: 
\code{\link{c19_eu_plots_save}()},
\code{\link{c19_heat_save}()},
\code{\link{c19_oblasts_save}()},
\code{\link{c19_r_plot_save}()},
\code{\link{c19_save_all}()},
\code{\link{c19_var_plot_save}()}
}
\concept{output funcs}
