% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendata_nsi.R
\name{c19_nsi_data}
\alias{c19_nsi_data}
\title{Provide access to NSI data}
\usage{
c19_nsi_data(reload = FALSE, redownload = FALSE)
}
\arguments{
\item{reload}{reload from disk}

\item{redownload}{refresh from internet}
}
\description{
Provide access to NSI data
}
\seealso{
Other data funcs: 
\code{\link{c19_bg_data}()},
\code{\link{c19_eu_data}()},
\code{\link{c19_reload}()}
}
\concept{data funcs}
