% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_r_plot.R
\name{c19_r_plot}
\alias{c19_r_plot}
\title{Plot R, incl. daily cases and 7d-rolling mean, 7d testing and positivity.}
\usage{
c19_r_plot(country_data = c19_bg_data())
}
\arguments{
\item{country_data}{country data}
}
\description{
R needs to be estimated first using \code{\link{c19_estimate_r}()}.
}
\seealso{
\code{\link{c19_estimate_r}()}

Other plot funcs: 
\code{\link{c19_deaths_age}()},
\code{\link{c19_deaths_factor}()},
\code{\link{c19_deaths_map}()},
\code{\link{c19_deaths_total}()},
\code{\link{c19_eu_weekly}()},
\code{\link{c19_heat}()},
\code{\link{c19_oblasts}()},
\code{\link{c19_var_plot}()}
}
\concept{plot funcs}
