% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c19_reload}
\alias{c19_reload}
\title{Reload all data from disk (or from the Internet).}
\usage{
c19_reload(redownload = FALSE)
}
\arguments{
\item{redownload}{whether to refresh all datasets from the Internet.}
}
\description{
Reload all data from disk (or from the Internet).
}
\examples{
\dontrun{
c19_reload()
c19_reload(redownload = TRUE)
}
}
\seealso{
Other data funcs: 
\code{\link{c19_bg_data}()},
\code{\link{c19_eu_data}()}
}
\concept{data funcs}
