% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{c19_save_all}
\alias{c19_save_all}
\title{Calls output funcs with some default params.}
\usage{
c19_save_all(
  file_ext = ".svg",
  dpi = 125,
  funs = c("c19_var_plot_save", "c19_heat_save", "c19_oblasts_save",
    "c19_eu_plots_save", "c19_r_plot_save"),
  dl = FALSE,
  rl = FALSE,
  ...
)
}
\arguments{
\item{file_ext}{extension (e.g. ".svg", ".png", ".jpg")}

\item{dpi}{dots per inch}

\item{funs}{functions to call (default: all)}

\item{dl}{whether to refresh all data sets from the Internet}

\item{rl}{whether to refresh all data sets from disk}

\item{...}{passed to individual save functions}
}
\description{
Output files will be located in c19bg/plots/save under current working
directory. Downloaded and calculated data will be stored in c19bg/data.
}
\examples{
\dontrun{
c19_save_all() # default is ".svg"
c19_save_all(file_ext = ".png")
c19_save_all(file_ext = ".jpg", quality = 100)
c19_save_all(file_ext = ".jpg", w = 10, h = 10, dpi = 200) #2000x2000px
c19_save_all(file_ext = ".png", dpi = 300)
}
}
\seealso{
Other output funcs: 
\code{\link{c19_eu_plots_save}()},
\code{\link{c19_heat_save}()},
\code{\link{c19_inkmagick}()},
\code{\link{c19_oblasts_save}()},
\code{\link{c19_r_plot_save}()},
\code{\link{c19_var_plot_save}()}
}
\concept{output funcs}
