% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_eu_plots.R
\name{c19_deaths_map}
\alias{c19_deaths_map}
\title{map plot                                                                     #}
\usage{
c19_deaths_map(eu_data = c19_eu_data())
}
\description{
map plot                                                                     #
}
