% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_eu_plots.R
\name{c19_deaths_total}
\alias{c19_deaths_total}
\title{contry totals plot (argument: country code, e.g. "BG", "UK", "EL")           #}
\usage{
c19_deaths_total(country_code, eu_data = c19_eu_data())
}
\description{
contry totals plot (argument: country code, e.g. "BG", "UK", "EL")           #
}
