% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_estimate_r.R
\name{c19_estimate_r}
\alias{c19_estimate_r}
\title{Estimate reproductive number for Bulgaria.}
\usage{
c19_estimate_r(country_data = c19_bg_data())
}
\arguments{
\item{country_data}{country data}
}
\description{
Uses package EpiEstim <\url{https://github.com/mrc-ide/EpiEstim}> by Cori et
al. <\url{https://doi.org/10.1093/aje/kwt133}> and generation times per
Ferretti et al. <\url{https://doi.org/10.1101/2020.09.04.20188516}>. Explores
a range of generation time distributions that are likely for Bulgaria. See
longer explanation in repo wiki:
<\url{https://github.com/StanTraykov/c19bg/wiki}>.
}
\details{
Also stores a hash of the input data (.Rhash in data dir) and skips
calculation if hash matches and estR.csv is already present.
}
\examples{
\dontrun{
c19_estimate_r() # save to csv file
result <- c19_estimate_r() # also save to variable
}
}
\seealso{
\code{\link{c19_r_plot}()}
}
