% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_eu_plots.R
\name{c19_eu_weekly}
\alias{c19_eu_weekly}
\title{Weekly plot using ECDC/EUROSTAT data.}
\usage{
c19_eu_weekly(
  indicator,
  continents = c("Asia", "Africa", "Europe", "Oceania", "America"),
  highlight = "BG",
  top_n = 20,
  lower_x = 10,
  lower_y = NA,
  label_fun = round,
  axis_labels = scales::label_number(),
  eu_data = c19_eu_data()
)
}
\arguments{
\item{indicator}{one of: "r14_cases", "r14_deaths", "em_1m", hosp_1m",
"positivity", "tests_100k"}

\item{continents}{continents to include (only some stats available outside
Europe/EU+)}

\item{highlight}{country to highlight (EU 2-ltr code, e.g. "EL" for Greece)}

\item{top_n}{number of lines to label}

\item{lower_x}{week axis limit (default NA = show all data)}

\item{lower_y}{indicator axis limit (default NA = show all data)}

\item{label_fun}{function to apply to line labels (default is to round)}

\item{axis_labels}{axis labels, e.g. scales::label_number(),
scales::label_percent()}

\item{eu_data}{eu data}
}
\description{
Weekly plot using ECDC/EUROSTAT data.
}
\examples{
\dontrun{
c19_eu_weekly(indicator = "r14_cases", lower_y = 0)
c19_eu_weekly(indicator = "positivity",
              top_n = 100,
              label_fun = function(x) sprintf("\%.1f\%\%", 100 * x),
              axis_labels = scales::label_percent())
}
}
\seealso{
Other plot funcs: 
\code{\link{c19_deaths_age}()},
\code{\link{c19_deaths_factor}()},
\code{\link{c19_deaths_map}()},
\code{\link{c19_deaths_total}()},
\code{\link{c19_heat}()},
\code{\link{c19_oblasts}()},
\code{\link{c19_r_plot}()},
\code{\link{c19_var_plot}()}
}
\concept{plot funcs}
