% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_heat.R
\name{c19_heat}
\alias{c19_heat}
\title{Plot incidence/100K heat map along age bands.}
\usage{
c19_heat(
  wday = "Monday",
  wrate = FALSE,
  first_wk = 1,
  country_data = c19_bg_data(),
  nsi_data = c19_nsi_data()
)
}
\arguments{
\item{wday}{weekday to right-align sums to ("Monday", "Tuesday", "Today",
etc.)}

\item{wrate}{whether to plot weekly growth rates (percentages) instead of
incidence}

\item{first_wk}{first week in 2020 to plot}

\item{country_data}{country data}

\item{nsi_data}{nsi data}
}
\description{
Plot incidence/100K heat map along age bands.
}
\seealso{
Other plot funcs: 
\code{\link{c19_deaths_age}()},
\code{\link{c19_deaths_factor}()},
\code{\link{c19_deaths_map}()},
\code{\link{c19_deaths_total}()},
\code{\link{c19_eu_weekly}()},
\code{\link{c19_oblasts}()},
\code{\link{c19_r_plot}()},
\code{\link{c19_var_plot}()}
}
\concept{plot funcs}
