% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_oblasts.R
\name{c19_oblasts}
\alias{c19_oblasts}
\title{Plot oblast incidence (per 100K or raw), faceted (map) or lumped ("spaghetti
plot").}
\usage{
c19_oblasts(incid_100k, facet = TRUE, country_data = c19_bg_data())
}
\arguments{
\item{incid_100k}{whether to plot raw incidence (case counts) or per 100K}

\item{facet}{whether to facet plot (individual plots on BG map)}

\item{country_data}{country data}
}
\description{
Plot oblast incidence (per 100K or raw), faceted (map) or lumped ("spaghetti
plot").
}
\seealso{
Other plot funcs: 
\code{\link{c19_deaths_age}()},
\code{\link{c19_deaths_factor}()},
\code{\link{c19_deaths_map}()},
\code{\link{c19_deaths_total}()},
\code{\link{c19_eu_weekly}()},
\code{\link{c19_heat}()},
\code{\link{c19_r_plot}()},
\code{\link{c19_var_plot}()}
}
\concept{plot funcs}
