% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_oblasts.R
\name{c19_oblasts}
\alias{c19_oblasts}
\title{Oblasts incidence plot}
\usage{
c19_oblasts(incid_100k, facet = TRUE, country_data = c19_bg_data())
}
\arguments{
\item{incid_100k}{whether to plot raw incidence or per 100K}

\item{facet}{whether to facet plot}

\item{country_data}{country data (defaults to BG data)}
}
\description{
Oblasts incidence plot
}
