% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c19_var_plot.R
\name{c19_var_plot}
\alias{c19_var_plot}
\title{Various plots}
\usage{
c19_var_plot(
  chart,
  roll_func = NULL,
  roll_window = 7,
  roll_align = "right",
  roll_fill = NA,
  line_legend = NULL,
  country_data = c19_bg_data()
)
}
\arguments{
\item{chart}{one of: "casesdeaths" | "age" | "dis" | "cases" |
"hospitalized" | "positivity" | "pospcr" | "posag"}

\item{roll_func}{function to roll, e.g. mean (default: NULL)}

\item{roll_window}{window for rull func (default: 7)}

\item{roll_align}{rolling alignment (default: "right")}

\item{roll_fill}{fill missing values for rolling function (default: NA)}

\item{line_legend}{regex specifying lines that should be labelled in-plot}

\item{country_data}{data}
}
\description{
Various plots
}
